refactor: separate ViewportParser into parser.ts and export noramlly
same reason to commit 166393
